update the scripts to resolve #1
we can't do fully automatic update with the API now because it doesn't
return the url to the download links as it did in pervious version.
The Api only send the url via email so we have to check email before
we can download the files.

So I make the etl into 2 phases: first send queries to API, and then
set the url in script and run it again.

Still need some changes to make it play well with airflow server.